split up GameView.vue into different components
