+++ src/components/DrawingComponent.vue
@@ -1,0 +1,134 @@
+<template>
+    <div class="drawing-area"
+         @mousedown="startDrawing"
+         @mousemove="draw"
+         @mouseup="stopDrawing"
+         @mouseleave="stopDrawing">
+      <canvas ref="canvas" width="800" height="500"></canvas>
+      <div class="current-color" :style="{ backgroundColor: penColor }"></div>
+      <div v-if="currentWord">
+        <GameButtons 
+          :changeStrokeColor="changeStrokeColor"
+          :changeLineWidth="changeLineWidth"
+          :undoLastStroke="undoLastStroke"
+          :resetCanvas="resetCanvas"
+        />
+      </div>
+    </div>
+  </template>
+  
+  <script>
+  import GameButtons from '@/components/GameButtons.vue';
+  
+  export default {
+    name: 'DrawingComponent',
+    components: {
+      GameButtons
+    },
+    props: {
+      currentWord: {
+        type: String,
+        default: ''
+      }
+    },
+    data() {
+      return {
+        isDrawing: false,
+        lastX: 0,
+        lastY: 0,
+        penColor: '#000000',
+        lineWidth: 7,
+        strokes: [],
+        currentStroke: []
+      }
+    },
+    methods: {
+      startDrawing(event) {
+        this.isDrawing = true;
+        this.lastX = event.offsetX;
+        this.lastY = event.offsetY;
+        this.currentStroke = [];
+      },
+      draw(event) {
+        if (!this.isDrawing) return;
+        const ctx = this.$refs.canvas.getContext('2d');
+        ctx.strokeStyle = this.penColor;
+        ctx.lineWidth = this.lineWidth;
+        ctx.beginPath();
+        ctx.moveTo(this.lastX, this.lastY);
+        ctx.lineTo(event.offsetX, event.offsetY);
+        ctx.stroke();
+  
+        this.currentStroke.push({
+          x1: this.lastX,
+          y1: this.lastY,
+          x2: event.offsetX,
+          y2: event.offsetY,
+          color: this.penColor,
+          width: this.lineWidth
+        });
+  
+        this.lastX = event.offsetX;
+        this.lastY = event.offsetY;
+      },
+      stopDrawing() {
+        if (this.currentStroke.length > 0) {
+          this.strokes.push([...this.currentStroke]);
+          this.currentStroke = [];
+        }
+        this.isDrawing = false;
+      },
+      changeStrokeColor(color) {
+        this.penColor = color;
+      },
+      changeLineWidth(width) {
+        this.lineWidth = width;
+      },
+      undoLastStroke() {
+        if (this.strokes.length === 0) return;
+        this.strokes.pop();
+  
+        const canvas = this.$refs.canvas;
+        const ctx = canvas.getContext('2d');
+        ctx.clearRect(0, 0, canvas.width, canvas.height);
+  
+        this.strokes.forEach(stroke => {
+          stroke.forEach(segment => {
+            ctx.strokeStyle = segment.color;
+            ctx.lineWidth = segment.width;
+            ctx.beginPath();
+            ctx.moveTo(segment.x1, segment.y1);
+            ctx.lineTo(segment.x2, segment.y2);
+            ctx.stroke();
+          });
+        });
+      },
+      resetCanvas() {
+        const canvas = this.$refs.canvas;
+        const ctx = canvas.getContext('2d');
+        ctx.clearRect(0, 0, canvas.width, canvas.height);
+        this.strokes = [];
+      }
+    }
+  }
+  </script>
+  
+  <style scoped>
+  .drawing-area {
+    position: relative;
+    width: 800px;
+    height: 600px;
+    border: 1px solid #000;
+  }
+  canvas {
+    display: block;
+    background-color: white;
+  }
+  .current-color {
+    margin-top: 15px;
+    width: 50px;
+    height: 50px;
+    border: 2px solid #000;
+    border-radius: 5px;
+  }
+  </style>
