+++ src/components/LeaderboardComponent.vue
@@ -1,0 +1,28 @@
+<template>
+    <div class="leaderboard">
+      <p>Leaderboard</p><br>
+      <div class="participant-name" v-for="participant in participants" :key="participant.name">
+        {{ participant.name }}
+      </div>
+    </div>
+  </template>
+  
+  <script>
+  export default {
+    props: {
+      participants: {
+        type: Array,
+        required: true
+      }
+    }
+  }
+  </script>
+  
+  <style scoped>
+  .leaderboard {
+    border: 1px solid #000;
+  }
+  .participant-name {
+    margin-bottom: 5px;
+  }
+  </style>
